DataMapper refactoring
