+++ src/main/java/junitparams/mappers/CsvWithHeaderMapper.java
@@ -11,23 +11,9 @@ import java.util.*;
  * @author Pawel Lipinski
  * 
  */
-public class CsvWithHeaderMapper implements DataMapper {
-    public Object[] map(Reader reader) {
-        BufferedReader br = new BufferedReader(reader);
-        String line;
-        List<String> result = new LinkedList<String>();
-        try {
-            try {
-                br.readLine(); // skip the header
-                while ((line = br.readLine()) != null) {
-                    result.add(line);
-                }
-                return result.toArray();
-            } finally {
-                reader.close();
-            }
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
+public class CsvWithHeaderMapper extends BufferedReaderDataMapper {
+
+    public CsvWithHeaderMapper() {
+        super(1);
     }
 }
